[BUGFIX] home directory of users and groups are not mounted in file list
The patch fixed the mount of the user and group dirs configured by the
TYPO3 conf vars. There are checks in front of the addMount call to
prevent exceptions in the Backend. Furthermore the patch fix the mount
name for the user, so "USERID_USERNAME" and also "USERID" are checked now.

Example configuration with the leading storage Uid for the fileadmin:
$GLOBALS['TYPO3_CONF_VARS']['BE']['userHomePath'] = '1:/users/';
$GLOBALS['TYPO3_CONF_VARS']['BE']['groupHomePath'] = '1:/groups/';

The base folders and the specific user/group folder have to be exists and
will not be created by this functionility.

Resolves: #48643
Releases: 6.2
Change-Id: Iad6e45fac16df8d6ba1503dfc4770a68d842394d