[FEATURE] Improve BE container view helper
Improve the features provided by the BE container view helper
so that extensions don't need to provide their own so often.
Added features:

- load several JS and CSS files instead of just one
- load jQuery
- load some inline language labels

Note: the old way of loading a single JS and CSS file is preserved
for backwards compatibility, but marked as deprecated.

With this patch all BE modules in the Core can use the default
BE container view helper. The Extension Manager needs some special
initialization which is moved to a dedicated view helper.

The customized BE container view helpers are removed.

Resolves: #49749
Releases: 6.2
Change-Id: I236134451a31f08afe5886337cfb1144dac7ad90
Reviewed-on: https://review.typo3.org/22063
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Felix Kopp
Tested-by: Felix Kopp