[FEATURE] FormEngine: Possibility to add icons via pageTSconfig
If someone adds a new value/label pair for a select field in
TCEforms/FormEngine, it is not possible to also add an icon to
the new option.
Alternatively, existing values can be changed through the new
"altIcons" property that acts like the "altLabels" property.

The patch adds another subelement for TSconfig to also add an
icon, as well as a new property.

TCEFORM.pages.doktype.addItems {
	13 = My Label
	13.icon = sysext/t3skin/icons/gfx/i/pages.gif
}
TCEFORM.pages.doktype.altIcons {
	123 = EXT:myext/icon.gif
}

The icon path must be relative to the typo3/ directory.

Resolves: #35891
Releases: master
Documentation: #49752
Change-Id: Ic835b1fe350144bc3498eab9a19857e3f558088d
Reviewed-on: http://review.typo3.org/22067
Reviewed-by: Alessandro Bellafronte <alessandro@4eyes.ch>
Reviewed-by: Christoph Dörfel <christoph.doerfel@gmail.com>
Reviewed-by: Michel Georgy <michel@4eyes.ch>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>