[FEATURE] FormEngine: Possibility to add icons via pageTSconfig
If someone adds a new value/label pair for a select field in
TCEforms/FormEngine, it is not possible to also add an icon to
the new option.
Alternatively, existing values can be changed through the new
"altIcons" property that acts like the "altLabels" property.

The patch adds another subelement for TSconfig to also add an
icon, as well as a new property.

TCEFORM.pages.doktype.addItems {
	13 = My Label
	13.icon = sysext/t3skin/icons/gfx/i/pages.gif
}
TCEFORM.pages.doktype.altIcons {
	123 = EXT:myext/icon.gif
}

The icon path must be relative to the typo3/ directory.

Resolves: #35891
Releases: 6.2
Documentation: #49752
Change-Id: Ic835b1fe350144bc3498eab9a19857e3f558088d