[BUGFIX] Respect HTTP_X_FORWARDED_PROTO for SSL
The HTTP Header "X-Forwarded-Proto" is a defactostandard which
may be sent by Frontend-Proxies/Load-Balancers to let its
Backend-Server know which protocol was used to start the
original request.

SSL Requests can be terminated at a transparent Frontend-Proxy
(e.g. nginx) and passed to the Backend-Server via plain HTTP
in this case.

This is the second attemp to fix the issue. The original
Change was voted down for security considerations. Now there
is a flag which needs to be set to true manually within
Install Tool ("AcceptXForwardedProtoSSL").

This should only be set in case the HTTP Header is explicitly
set by the Frontend-Proxy and cannot be passed in by a client.

Change-Id: Ie6ab10f503737b2146032911b2587e1e96e9db0f