[BUGFIX] Page module: Allow to paste in empty columns
Since quite a long time it was not possible to paste a content
element into an empty column of the page module. This patch
introduces an icon in the top right corner of each column if
a tt_content element is on the clipboard (Default pad). When
the icon gets clicked the content elements from the clipboard
get moved/copied into the selected column/language.

Change-Id: I635bb584d5e03b764bd5df5189555fb9b34ab926
Resolves: #44002, #35980, #15080, #21622
Releases: 6.1, 6.2