[BUGFIX] Page module: Allow to paste in empty columns
Since quite a long time it was not possible to paste a content
element into an empty column of the page module. This patch
enables the clickMenu (context menu) for the header of a content
column. From this context menu which can get activated via a
mouse click (right or left) the "paste" option can get choosen
to move/copy the current clipboard element into the selected
column/language.

Change-Id: Ib931dfed4cbd1678e1f1d2af9b1e7b6079b09d71
Resolves: #44002, #35980, #15080, #21622
Releases: 6.1, 6.2