[BUGFIX] Indexing files does not store the language
While indexing externals (e.g. PDF files) the language of the page
being rendered is not stored with the indexed file content. Thus
the file is shown/not shown in the wrong situations.

See detailed description in bug-report regarding the "situations".

This is a manual back-port from the merged change in master,
just doing the same one-line addition.

Change-Id: Id949648619adadc0727eddb4e42a370a875bae5b
Resolves: #46261
Releases: 4.7, 6.0, 6.1, 6.2