[BUGFIX] Faulty layout if row details are too long
If the row details in in the workspaces difference view are too
long, e.g. a string that does not contain white-spaces like e.g.
a filename, the layout looks broken.
The solution is to introduce a new classname for the content
details and to force word breaking via CSS.

Fixes: #50114
Releases: 6.0, 6.1, 6.2
Change-Id: I9a768aed1c9bb4ebdfea5ba79083b7e5dce262d9
Reviewed-on: https://review.typo3.org/22387
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters