[WIP][TASK] Use alternative selector for many workspaces
The toolbar of the workspace module visualizes each workspace
in a separate tab. On having many workspaces, the toolbar is
extended to multiple lines which actually reduces the available
viewport of the grid panel.

This patch introduces the native ExtJS TabPanel and an
alternative ComboBox to directly select the accordant
workspace if the available toolbar width is exceeded.

Change-Id: I85c8b8134dfe27a4dc32c5d70a158cca41763c41