[FEATURE] uriBuilder should honor POST arguments for addQueryString
Currently the addQueryString option in the uriBuilder only supports
arguments which are in the query string. Therefore, the pagination
widget does not work with filtering post requests.

Solution: Add addQueryStringMethod (already supported by typolink
configuration) as property of uriBuilder and as argument of the
following fluid view helpers:
* link.action
* link.page
* uri.action
* uri.page
* widget.link
* widget.uri
* widget.pagination
Add addQueryStringMethod as configuration option of the paginate widget.

Fixes: #35281
Fixes: #11441
Releases: 6.2
Change-Id: I8f2963fa7467dffcf3b9535d2a41820de8f9930a