[BUGFIX] Page tree filtering broken in IE7 & IE8
The page tree doesn't show the result when using the filter
in the page tree in IE7 & IE8. ExtJS somehow breaks silently
when trying to use a non-tag (text) as a drag handle. This fix
works with the original change that introduced this, meaning it
doesn't break the functionality of dragging of locked records.

Change-Id: I3f33fb3e4f6bd7622f502a9265897fee9d3e4ed3
Resolves: #25327
Resolves: #37026
Related: #M17952
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Reviewed-on: https://review.typo3.org/23339
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes