[BUGFIX] Object's children are not translated anymore
As a result from the patch to issue #47192 the translation of an object's
child objects does not work anymore.

I have an object *Journey* with 1:n *Destination*. I have a list of
journeys where I show the destination's name. In the default language
(German) everything is fine. If I open the page with language id 1
(English) the labels and contents are shown translated but not the
destinations' names which are translated in the BE of course. The problem
here is, that when mapping the properties to child objects the query
setting "respectSysLanguage" is set to false in
DataMapper::getPreparedQuery, which is absolutely right.

The error happens in Typo3DBBackend::doLanguageAndWorkspaceOverlay, when
it comes to overlaying the current record. Due to the patch the overlay
is only done if "respectSysLanguage" ist true, which means, it is never
done for child objects. As it is necessary to set "respectSysLanguage" to
false and "returnRawQueryResult" to true to achieve the behavior
described in #47192 I suggest to check for the second one also: Do the
overlay if either "respectSysLanguage" is true *OR*
"returnRawQueryResult" is false. That does not destroy returning a raw
record, as I have tested.

Resolves: #50417
Releases: 6.2, 6.1, 6.0
Change-Id: I1015da75c18373d219f16fa7fe8d0c6df374dc76