[BUGFIX] FAL does not copy subfolders cleanly
If full structures with sub folders are copied around in FAL
local driver, sub folders are not correctly created in the
target directory.

This is detected in unit tests after recent vfsStream updates in
ext:phpunit.

The patch adds a flag to the Iterator code to also include
directories to fix the issue.

Additionally, the existing two unit tests also tested multiple other
methods of the local driver and are functional tests. The patch moves
those two tests to the functional section and adds three simple unit
tests with correct mocking instead.

Change-Id: I30d3c8bd5c6538b3fe23405d97bcddff6c1ff405
Releases: 6.2, 6.1, 6.0
Resolves: #50472