[BUGFIX] "New page" wizard discloses existence of pages outside DB mount
When creating a new page inside the top level of a DB mount which is
only a sub tree, the pages up and down from the DB mount root will be
displayed in the position selector if the logged in user has read
permissions for these pages. This is unwanted information disclosure as
the permissions should not matter for pages which are outside the DB
mount.

Resolves: #18797
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I98008bc7f4308c9fb32dae645325e7cb1b44e413