[BUGFIX] Fix fatal error in extension manager
TER extensions which were uploaded just recently contain a change in
ext_emconf.php. This change conflicts with older TYPO3 releases:

If the extension has no conflicts, the constraint is written as an empty
string although it is expected to be an array in any case.

This fix adds a check to work around the problem.

Change-Id: Ibb361f82c39d99089ec6455935ddc3e775cebbc7
Resolves: #46642
Releases: 4.7, 4.5
Reviewed-on: https://review.typo3.org/22657
Reviewed-by: Tomita Militaru
Tested-by: Tomita Militaru
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters