[FEATURE] Pre-fill OpenID backend login field
After successful login, a cookie gets set in the user's browser
containing the OpenID url. Whenever the user visits the backend login
form, this URL is pre-filled into the username field.

When the user switches between normal login and openid login,
the pre-filled openid will be removed and added.

The user may choose to store or not store the OpenID. If he chooses not
to store it, it will not be pre-filled.

Resolves: #50545
Releases: 6.2
Depends: Id31238760fe4e58e2a823beddaeb454ba28d59be
Change-Id: I6cfdb70e8ae9b5afdfd6ecb1b9ee557e3c2959ce