[TASK] Travis setup without Introduction package
In travis, the introduction package is used to import the database
schema to successfully run the unit tests.
This is now substituted with a small php based script in the
travis integration repository, that basically imports the base schema
by running code similar to the according step in the installation
process. Cloning the introduction package is no longer needed.
Travis will now for automatically adapt to database schema changes
defined in ext_tables.sql scripts and without changing foreign
repositories anymore.

Change-Id: If246650acfabde119d4e02c78d5449ab8d44feef
Resolves: #50571
Releases: 6.2