[FEATURE] Use salted Install Tool password
To enhanced the security change Install Tool password from md5 hash to
salted. This means we have to change the default password for new
installations as well as the comparison value for different tests on the
Install Tool password. Besides there is still the possibility to use
md5 hashed passwords which gets converted on first usage. This should
help to set a new Install Tool password if it got lost.

Resolves: #50613
Releases: 6.2
Change-Id: If02a43780c9c819ebd6da7cbf0acad305f805330