[BUGFIX] Fix failing test
With #50803 the usage of the constant TYPO3_MODE was introduced. This
leads to the failing test, after the constant can not be redefined
for test purposes.
To work around this, a method parameter was introduced, that allows the
definition of the environment mode for the methode under test. So the
test works again without changes.
As a sideeffect code can now define at runtime for what environment mode
the method will prepare the result.

Resolves: #50809
Relates: #50803
Releases: 6.2
Change-Id: I7319d8ccef530feff69a7f9885df24dba0beed4d