[BUGFIX] Fix failing test
With #50803 the usage of the constant TYPO3_MODE was introduced. This
leads to the failing test, after the constant can not be redefined
for test purposes.
To work around this, a getter was introduced, that allows to mock the
definition of the environment mode for the methode under test.

Resolves: #50809
Relates: #50803
Releases: 6.2, 6.1, 6.0
Change-Id: I7319d8ccef530feff69a7f9885df24dba0beed4d