[BUGFIX] Make sure in unit tests the singletons are backup at first
If in unit tests the singleton management is changed there is a local
copy of the initial stage. If the backup contains classes which were
created during setup method after finishing the test the internal
__destruct method is called. This is why it is necessary to store the
singleton classes before any other initialization otherwise the
construct function is not executed for other tests and properties are
wrong or missing.

Resolves: #50812
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I7788a5788aca6777040b777aeceb0558f96022b2