[BUGFIX] Use proper implementation of a wordwrap
The page module uses a custom implementation of a wordwrap
as this was not supported by the browser in those earlier days.

The big drawback is that there is no UTF8 support for this which
leads to empty preview boxes when e.g. using Japanese.

The 6.1 version of this patch changes the wordWrapper() usage only
and does not deprecate the method.

Change-Id: I7caab8e14a9a0b7ccbd38c692b0cf98635967567
Resolves: #15797
Releases: 6.2, 6.1