[BUGFIX] Add missing API method FileInterface::getNameWithoutExtension
This method was supposed to be in the interface, but due to a syntax
error in the closing comment tag, it was not (and was not detected as a
PHP syntax error).

This finally adds this method and implements it also in the class
FileReference which hadn't had this method before but also implements
this interface.

Resolves: #51509
Releases: 6.2, 6.1, 6.0
Change-Id: I8ee1ecc155e9cfaeec7ac0ac7086f75df13f4c34