[BUGFIX] Failing Resource\FactoryTest on Windows systems
Two tests in class TYPO3\CMS\Core\Tests\Unit\Resource\FactoryTest fail
with warnings. This is because of two single issues.

First of all the test depends on a folder which (might) belong to an own
storage other than storage 0 (fileadmin). This patch changes fileadmin
to the typo3 folder as this one should be not included in any storage.

On the other hand on Windows systems a full path contains always a colon
which breaks any storage check. To solve this we have to fix any
absolute path beforehand as it is done in the sub functions as well.

Resolves: #50843
Releases: 6.2, 6.1, 6.0
Change-Id: I950a3c81222155da403ca1eb7b920e8682033450
Reviewed-on: https://review.typo3.org/22912
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes