[BUGFIX] Test in Configuration\ConfigurationManagerTest fails in backend
When running the tests in the PHPUnit backend, the path of the current
working directory is typo3/ (where mod.php is included). Therefore the
test canWriteConfigurationReturnsTrueIfDirectoryAndFilesAreWritable
fails as it deals with paths relative to the root directory. We have to
use absolute paths to ensure files in typo3temp folder exist.

Resolves: #50845
Releases: 6.2, 6.1
Change-Id: I7859978c31cf523c8803316fdac8c067d00042cc
Reviewed-on: https://review.typo3.org/22898
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind