[TASK] Introduce AbstractHierarchicalFilesystemDriver
There are many tasks which apply to all drivers working on an
Hierarchical FileSystem (NFS, WebDav, LocalDriver, ....).
When it comes to security we dont want custom drivers to implement
that all on their own, but provide it by the System.

Introducing AbstractHierarchicalFilesystemDriver enables us to
do so, without breaking compatibility (Base class, interface
and Classname stay the same).

Releases: 6.2, 6.1, 6.0
Resolves: #50867
Change-Id: Ib1b45c6ad384f82f7a5138372ed91269b8db38c2