[BUGFIX] Correctly set user storage permissions
With FAL a new way to define file permissions
(for storages) has been introduced. You can set
these in User TsConfig instead of doing it
in the user record.

There is a backwards compatibility layer, which should
set the permissions according to the database entry
in case no permissions have been set in TsConfig.
This however does not work, because of a wrong check
for existence of UserTsConfig properties.

Instead of checking for an array (which is always the case),
we have to check if the array is not empty.

Releases: 6.0, 6.1, 6.2
Resolves: #50872
Change-Id: I0ec30a1ae96c4b6916a3e46c7222f832c44f603e
Reviewed-on: https://review.typo3.org/22929
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel