[BUGFIX] Tests in Localization\Parser\LocallangXmlParserTest fail
If the system is already translated into French two tests in
Localization\Parser\LocallangXmlParserTest fail. This is because another
localization file is taken where test strings are not included.

As the tested functions can't work system independent, the test has to
make sure no system dependencies are used. Therefore the patch
introduces a new language "md5" to prove the workflow.

Resolves: #50870
Releases: 6.2, 6.1, 6.0
Change-Id: I64e8864a7d746c4f5c278f5e67d8dae068d8f920