[BUGFIX] Fix file permission methods in BackendUserAuthentication
Take default TSConfig file permissions into account:

File permissions configured in User TSConfig,
are not taken into account, because the properties
from the getTSConfig method are returned in a
'properties' key of an array but the top level
array is used. Use getTSConfigProp instead.

Fix the bit wise check for old file permissions:

XOR the value is wrong because it gives the wrong
result if more bits are set. Use AND operation
to properly check if a specific bit is not set.

Check if the user is admin in getFilePermissionsForStorage()

Add tests to confirm the desired behaviour.

Releases: 6.0, 6.1, 6.2
Resolves: #51004
Change-Id: Ia5d6fa1cb47a74306fe5465a0e70c2f2aea2a4b8
Reviewed-on: https://review.typo3.org/23111
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel