[BUGFIX] Fix file permission methods in BackendUserAuthentication
Take default TSConfig file permissions into account:

File permissions configured in User TSConfig,
are not taken into account, because the properties
from the getTSConfig method are returned in a
'properties' key of an array but the top level
array is used. Use getTSConfigProp instead.

Fix the bit wise check for old file permissions:

XOR the value is wrong because it gives the wrong
result if more bits are set. Use AND operation
to properly check if a specific bit is not set.

Check if the user is admin in getFilePermissionsForStorage()

Add tests to confirm the desired behaviour.

Releases: 6.0, 6.1, 6.2
Resolves: #51004
Change-Id: I78c5fa13f2110e6044c0a81e2e7e4f3298766226
Reviewed-on: https://review.typo3.org/23125
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel