[BUGFIX] Create workspace label placeholder that matches field conditions
If a workspace placeholder record is generated the field, that
is configured as label for a table, is filled with a placeholder.
This is a hardcoded string.
If database field does not allow strings (integer, double...)
a TYPO3 warning is displayed. "... Propably value mismatch ...".
This message is very confusing because the user does not know
where it comes from.

The label should be evaluated against the TCA eval configuration
before it is written to the database.

Resolves: #17551, #31757
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I18261359550dcaddaa9bd9ca0dd77f8300e81da9