[BUGFIX] Take into account all file and folder permissions
When handling the old file permission settings from
user or user group records, all current FAL user
permissions should be set accordingly, to not allow
file actions with FAL which were not allowed before.

In particular no write / change actions should be
allowed if no permissions are granted in the records.

Because of #51007 we also need to test for
$defaultPermissions['properties'] instead of
$defaultPermissions as we now always get an array
with 'properties' key set.

Releases: 6.0, 6.1, 6.2
Resolves: #51005
Related: #51007

Change-Id: I873f23ed02428842621c8256df5792e54abbfe2c