[BUGFIX] Fix inconsistencies in getTSConfig in BackenuserAuth
BackenuserAuthentication::getTSConfig() behaves
inconsistently when an object string is given with
multiple sections but no result in the actual TSConfig
of the user. Other than noted in the method declaration,
the method returns an empty array instead of an array
where the 'properties' key is null.

Additionally we fix a bug which caused wrong results
if "0" was used as a key in configuration and accessed
with something like "permissions.file.0"

Resolves: #51007
Releases: 6.0, 6.1, 6.2
Change-Id: I80fc7398c1955ff77d052377e3ba81bb8aea01cf
Reviewed-on: https://review.typo3.org/23032
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Markus Klein
Tested-by: Markus Klein