[BUGFIX] Fix inconsistencies in getTSConfig in BackenuserAuth
BackenuserAuthentication::getTSConfig() behaves
inconsistently when an object string is given with
multiple sections but no result in the actual TSConfig
of the user. Other than noted in the method declaration,
the method returns an empty array instead of an array
where the 'properties' key is null.

Additionally we fix a bug which caused wrong results
if "0" was used as a key in configuration and accessed
with something like "permissions.file.0"

Resolves: #51007
Releases: 6.0, 6.1, 6.2
Change-Id: Ie3daa0abc04a927186e9e9f3a88875086b606f93
Reviewed-on: https://review.typo3.org/23123
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel