[BUGFIX] Fields of type file_reference are not properly indexed
sys_refindex entries for TCA-fields of type group->file_reference
are not correct. The fields ref_table and ref_uid are not filled.
This results in a wrong number of references in the Ref column of
the Filelist module and prevents a warning that a resource is still
in use when deleting it.

Resolves: #49538
Releases: 6.2, 6.1, 6.0
Change-Id: I1c44fc98b7ceefd6247ec372e5b28f8682a47bf1
Reviewed-on: https://review.typo3.org/24449
Reviewed-by: Markus Klein
Tested-by: Markus Klein