[BUGFIX] DataHandler::clear_cache() does not consider workspaces
DataHandler::clear_cache() is called with a table and a uid that
has been modified to trigger the clear cache commands. To find
siblings, children, parents, etc. the pid value is directly used
to find the affected pages. However, in a workspace context, the
pid is always "-1" which leads to a selection of all page
records of all workspaces. This amount is used to flush all
defined caches by a given tag name and might result in execution
time outs.

Caches in the front-end use the live id of the page to be shown,
thus DataHandler first has to resolve the live page id to clear
the caches.

Fixes: #51051
Releases: 6.0, 6.1, 6.2
Change-Id: I03987bbd8c5d3526fb4dfda75b0dd0316fe1d836