[WIP][FEATURE] Unify translation handling for pages table
Table 'pages' currently uses an own table 'pages_language_overlay' to
store translated records. This is unique in the Core and done by
several hardcoded conditions through the Core.
In order to unify the handling to ease functional testing for trans-
lation and workspaces functionality the unique handling will be
removed.

Resolves: #51139
Releases: 6.2
Change-Id: I7b4140d701eff78b44cdbea32765726c7f31fd77