[FEATURE] Allow signalSlots to modify arguments
To allow signalSlots to modify the given arguments,
even if not an object (which could be modified by
reference anyway), the signalSlot should be able
to return the arguments.

This will make the signal slot handling diverge
from Flow and violates or at least bends the
signal slot pattern.

To make this clear, we should rename the
signal slot dispatcher in an upcoming change.

Resolves: #51138
Releases: 6.2
Change-Id: I36928adceed672580c76a68891dfb825e43a5a9e