[FEATURE] move configuration of returnRawQueryResult
In motion to streamline the persistence configuration we will
get rid of the Typo3QuerySettings. Especially for
returnRawQueryResult it does not make sense to have a
configuration for that.

We moved it to be a method argument, so some can still decide
on a per-query-basis, what the query is supposed to return.

Resolves: #51145
Releases: 6.2
Change-Id: I948267149b66c350b271d3c98c5cbe3e268d087e