[BUGFIX] FlashMessageQueue PHP strict warning
FlashMessageQueue exends \SplQueue and overwrites enqueue($value).
In \SplQueue enqueue() has no type hint on $value, but
FlashMessageQueue adds a type hint to FlashMessage. This raises
a PHP strict warning.

Solution is to remove the type hint but at an exception if given
argument is not an instance of FlashMessage.

Change-Id: I882aab33a55d84de6614a45c6cc045223adba0cd
Resolves: #51159
Releases: 6.2, 6.1