[BUGFIX] Catch all errors while starting installer
While initializing the installer it can happen that
Exceptions are thrown. As there is no ExceptionHandler
initialized yet, we get a white page and the Fatal error
will be logged.

Mostly it is the problem that typo3conf and/or typo3temp
isn't writeable. But it might something else which throws
an exception.

This patch catches the exception in the initializing
process and shows a page in the layout of the installer.
It shows the exception message, a first hint for the maybe
write permissions problem and a link with the exception number.

Not needed for 6.2 as there the installer is completely
rewritten.

Resolves: #11771
Releases: 6.1, 6.0
Change-Id: Ib99860054b727907a9a84bdfd3e35fe3f2eebe00