[FEATURE] Force saltedpasswords for Backend
This patch adds saltedpasswords as a required system extension. The
possibility to disable backend hashing is removed and during the install
tool steps the environment settings are checked. Rsaauth isn't
forced anymore but it's possible to use saltedpasswords with an
unencrypted transfer mode. Besides rsaauth is added to the list of
recommended extensions during update process.

Resolves: #51352
Releases: 6.2
Change-Id: Ifd3b9f195101bcdb083b3bf9db4a74c812f0a709