[BUGFIX] Install tool: Check for broken ext uninstalls correct ext
"Check for broken extensions" in important actions of the install
tool renders uninstall buttons for every broken extension it found.

Currently only the last extension is uninstalled, no matter which
button is used. The patch fixes this by adding a class to identify
the extension in question to each button and adapts the javascript.

Resolves: #51368
Releases: 6.2
Change-Id: I7b754aca58e262dbd726452f816fe4dfaa6cf54d
Reviewed-on: https://review.typo3.org/23336
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Stefan Neufeind
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters