[FEATURE] Wrong HTTP headers when trying to access protected pages
When trying to access a login protected page while not being logged in
it may happen that a wrong header is sent.
The install tool allows to set a header for the page not found
handling: [FE][pageNotFound_handling_statheader]
The default value for this setting is "HTTP/1.0 404 Not Found"
In a case where access is denied because the user is not logged in,
this is the wrong header, it should be 403 instead. Therefore this patch
implements a new configuration [FE][pageForbidden_handling_statheader]

Resolves: #23178
Releases: 6.2
Change-Id: I6558fedc21ef3dfc203bccf2d39f8522f90109c6