[TASK] prevent too long functional test database names
The functional tests frameworks creates a new database.
The database name is generated by the original database name
plus a string and a hash.

If the database name length exceeds the maximum length
for database names in mysql an exception  is now thrown.

The generated database name has also been shortened

Resolves: #51383
Releases: 6.2
Change-Id: I5817e440a3dcba7e53e2a27dca4f7bfc679334f5