[FEATURE] Install Tool: Check for broken ext indicates loading
"Check for broken extensions" in important actions of the install
tool checks for broken extenions over an Ajax-Request.

Currently, the user clicks the button to check for broken extensions
and doesn't get any feedback whether there is something going on or not.
This patch implements a loading-message with a Spinner that is displayed
as soon as the button to check for broken extensions is clicked.
Furthermore the button is hidden as long as the request is beign executed
and displayed again if there is anything more to check (e.g. broken
extensions are found).

Resolves: #51379
Release: 6.2
Change-Id: I13e5024a1e2666e54e782ff1bab9418ba8b180f6