[FEATURE] Configuration checks for resource storage
The configuration for a storage can have errors in quite a few places.
To help admins while creating a storage, a new field is added in the
forms for storage records. A custom user function renders hints,
warnings and errors there.

The checks implemented by this patch are: case-sensitive file system and
"is this storage part of another storage or does it contain another
storage?" (see #50796).

The configuration check could also be opened for custom methods
registered e.g. by extensions providing custom drivers.

Change-Id: If94f457058fe4a5ebdd20e870e9b56e075ecdb23
Related: #50796
Releases: 6.2