[TASK] Remove misleading icon from buttons in install tool
Nearly all buttons in the install tool currently show a green checkmark
icon. This is misleading, as in most cases nothing has been checked.
The button simply leads to the next step or executes the desired action.

With removing the icon, it will be more clearer that clicking a button
will just perform the intended action, without any check being
performed beforehand that this action will actually succeed.

Change-Id: If00d4ccde24a201f552ead6b9d8991c8e58b473f
Resolves: #51407
Releases: 6.2